* Bring up Windows port to Google's v1.6.7 r1.
* Add new source to project files (both vcxproj and CMake).
* Bump SOVERSION to 10 in all CMakeLists.
* Better workaround for a template bug in MSVC.
* Remove unneeded templace specializations for the new Power()
  function that made the MSVC compiler icky.

+++ src/include/fst/equivalent.h
@@ -112,7 +112,7 @@ constexpr
 //        by 4 by all practical purposes.
 template <class Arc>
 bool Equivalent(const Fst<Arc> &fst1, const Fst<Arc> &fst2,
-                double delta = kDelta, bool *error = nullptr) {
+                float delta = kDelta, bool *error = nullptr) {
   using Weight = typename Arc::Weight;
   if (error) *error = false;
   // Check that the symbol table are compatible.
